+++ packages/fast-html/test/fixtures/event/event.spec.ts
@@ -7,35 +7,25 @@ test.describe("f-template", async () => {
         const customElement = page.locator("test-element");
 
         let message;
-        page.on("console", msg => message = msg.text());
+        page.on("console", msg => (message = msg.text()));
 
         await customElement.locator("button").nth(0).click();
 
         expect(message).toEqual("no args");
     });
-    test("create an event attribute with an event argument", async ({ page }) => {
+    test("create an event attribute with an event argument (deprecated e)", async ({
+        page,
+    }) => {
         await page.goto("/fixtures/event/");
 
         const customElement = page.locator("test-element");
 
         let message;
-        page.on("console", msg => message = msg.text());
+        page.on("console", msg => (message = msg.text()));
 
         await customElement.locator("button").nth(1).click();
 
         expect(message).toEqual("click");
-    });
-    test("create an event attribute with an attribute argument", async ({ page }) => {
-        await page.goto("/fixtures/event/");
-
-        const customElement = page.locator("test-element");
-
-        let message;
-        page.on("console", msg => message = msg.text());
-
-        await customElement.locator("button").nth(2).click();
-
-        expect(message).toEqual("bar");
     });
     test("should properly bind events with `this`", async ({ page }) => {
         await page.goto("/fixtures/event/");
@@ -44,8 +34,58 @@ test.describe("f-template", async () => {
 
         await expect(customElement).toHaveJSProperty("foo", "bar");
 
-        await customElement.locator("button").nth(3).click();
+        await customElement.locator("button").nth(2).click();
 
         await expect(customElement).toHaveJSProperty("foo", "modified-by-click");
+    });
+    test("create an event attribute with $e argument", async ({ page }) => {
+        await page.goto("/fixtures/event/");
+
+        const customElement = page.locator("test-element");
+
+        let message;
+        page.on("console", msg => (message = msg.text()));
+
+        await customElement.locator("button").nth(3).click();
+
+        expect(message).toEqual("click");
+    });
+    test("create an event attribute with $c (context) argument", async ({ page }) => {
+        await page.goto("/fixtures/event/");
+
+        const customElement = page.locator("test-element");
+
+        let message;
+        page.on("console", msg => (message = msg.text()));
+
+        await customElement.locator("button").nth(4).click();
+
+        expect(message).toEqual("click");
+    });
+    test("create an event attribute with multiple arguments ($e, $c)", async ({
+        page,
+    }) => {
+        await page.goto("/fixtures/event/");
+
+        const customElement = page.locator("test-element");
+
+        let message;
+        page.on("console", msg => (message = msg.text()));
+
+        await customElement.locator("button").nth(5).click();
+
+        expect(message).toEqual("click,click");
+    });
+    test("create an event attribute with $c.event argument", async ({ page }) => {
+        await page.goto("/fixtures/event/");
+
+        const customElement = page.locator("test-element");
+
+        let message;
+        page.on("console", msg => (message = msg.text()));
+
+        await customElement.locator("button").nth(6).click();
+
+        expect(message).toEqual("click");
     });
 });